Added standard.html template, updated css
The standard.html template serves as a base to all pages except
the index page. It inherits from base and the other pages inherit
from it. This was done as the pages were all sharing the same
starting html and now it is all defined in one spot instead of
many.